requirements.txt
selenium>=4.15.2
python-dotenv>=1.0.0
langchain>=0.1.0
langchain-google-genai>=0.0.5
langgraph>=0.0.15
pydantic>=2.5.2
typing-extensions>=4.8.0
python-json-logger>=2.0.7
requests>=2.31.0
beautifulsoup4>=4.12.2
webdriver-manager>=4.0.1 
google-generativeai==0.3.2 